Moved webpack common into modules folder
