+++ src/renderer/modules/common/constants.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unnecessary-type-assertion */
-import { filters, getExportsForProps, waitForModule } from "..";
+import { filters, getExportsForProps, waitForModule } from "../webpack";
 
 export const raw = await waitForModule(filters.bySource("BASE_URL:"));
