+++ src/renderer/modules/common/contextMenu.ts
@@ -1,5 +1,5 @@
 import type { ObjectExports } from "../../../../types";
-import { filters, getFunctionBySource, waitForModule } from "..";
+import { filters, getFunctionBySource, waitForModule } from "../webpack";
 
 export interface ContextMenu {
   close: () => void;
