+++ src/renderer/modules/common/flux.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type { RawModule, ReactComponent } from "../../../../types";
 import type { FluxDispatcher as Dispatcher } from "./fluxDispatcher";
-import { filters, waitForModule } from "..";
+import { filters, waitForModule } from "../webpack";
 
 type DispatchToken = string;
 type ActionType = string;
