+++ src/renderer/modules/common/hljs.ts
@@ -1,5 +1,5 @@
 import HighlightJS from "highlightjs";
-import { filters, waitForModule } from "..";
+import { filters, waitForModule } from "../webpack";
 
 export default await waitForModule<typeof HighlightJS>(
   filters.byProps("initHighlighting", "highlight"),
