+++ src/renderer/modules/common/lodash.ts
@@ -1,4 +1,4 @@
-import { filters, getExportsForProps, waitForModule } from "..";
+import { filters, getExportsForProps, waitForModule } from "../webpack";
 
 import type Lodash from "lodash";
