+++ src/renderer/modules/common/modal.ts
@@ -1,6 +1,6 @@
 import type { ObjectExports, RawModule } from "../../../../types/webpack";
 import type { ReactComponent } from "../../../../types/util";
-import { filters, getBySource, getFunctionBySource, waitForModule } from "..";
+import { filters, getBySource, getFunctionBySource, waitForModule } from "../webpack";
 
 enum ModalTransitionState {
   ENTERING,
