+++ src/renderer/modules/common/react-dom.ts
@@ -1,4 +1,4 @@
-import { filters, waitForModule } from "..";
+import { filters, waitForModule } from "../webpack";
 import type ReactDOM from "react-dom";
 
 export default await waitForModule<typeof ReactDOM>(filters.byProps("createPortal", "flushSync"));
