+++ src/renderer/modules/common/toast.ts
@@ -1,5 +1,5 @@
 import type { ObjectExports } from "../../../../types";
-import { filters, getFunctionBySource, waitForModule } from "..";
+import { filters, getFunctionBySource, waitForModule } from "../webpack";
 
 const Kind = {
   MESSAGE: 0,
