+++ src/renderer/modules/webpack.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/unified-signatures */
-import type { AnyFunction } from "../../../types/util";
+import type { AnyFunction } from "../../types/util";
 import type {
   Filter,
   GetModuleOptions,
@@ -18,7 +18,7 @@ import type {
   WebpackChunkGlobal,
   WebpackModule,
   WebpackRequire,
-} from "../../../types/webpack";
+} from "../../types/webpack";
 
 // Handlers
